refactor:ディレクトリ構成を機能種別ごとに変更
