+++ app/event/delete_token.js
